% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_arithmetic_mean.R
\name{fct_arithmetic_mean}
\alias{fct_arithmetic_mean}
\title{Calculate emissions based on arithmetic means}
\usage{
fct_arithmetic_mean(.checked_data)
}
\arguments{
\item{.checked_data}{The list returned by \code{fct_checkinput()}. Its
\code{template_version} element drives version-specific behaviour and its
\code{data} element supplies the \code{setup}, \code{time}, \code{area}
and \code{carbon} tables.}
}
\value{
A data frame with arithmetic mean of CO2 emissions for each land use
transition, REDD+ activity or emission reductions level.
}
\description{
TBD
}
\examples{
library(mocaredd)

path <- system.file("extdata/mocaredd-templatev2-simple.xlsx", package = "mocaredd.dev2")

checked <- fct_checkinput(.path = path)

res <- fct_arithmetic_mean(.checked_data = checked)
res$gg_emissions

}
