% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_data2.R
\name{fct_check_data2}
\alias{fct_check_data2}
\title{Check that the app's input data meet the template criteria.}
\usage{
fct_check_data2(.usr, .time, .ad, .cs)
}
\arguments{
\item{.usr}{User inputs' table for the shiny app (user_inputs)}

\item{.time}{the 'time' table from the tool input file (see template)}

\item{.ad}{Activity Data input table for the shiny app (AD_lu_transitions)}

\item{.cs}{Carbon Stock input table for the shiny app (c_stocks)}
}
\value{
A dataframe with TRUE or FALSE (TRUE if each check passes), and broad error locations
if FALSE.
}
\description{
Tests 6 types of conformity with the app template to avoid code break during
analysis: (1) column names, (2) tables' size, (3) data types,
(4) categories names, (5) unique IDs and (6) matching key variables between tables.
}
\examples{

path <- system.file("extdata/example1-4pools.xlsx", package = "mocaredd.dev2")
path <- system.file("extdata/mocaredd-template-v2-4pools.xlsx", package = "mocaredd.dev2")

tabs <- c("user", "time", "area", "carbon")
dat <- purrr::map(tabs , function(x){
  readxl::read_xlsx(path = path, sheet = x, na = "NA")
})
names(dat) <- tabs

fct_check_data2(.ad = ad, .cs = cs, .usr = usr, .time = time)

}
