% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_ER.R
\name{fct_combine_mcs_ER}
\alias{fct_combine_mcs_ER}
\title{Combine MCS of emissions from Reference and monitoring period into emission reductions}
\usage{
fct_combine_mcs_ER(.sim_ref, .sim_mon, .ad_annual)
}
\arguments{
\item{.sim_ref}{simulation aggregated to reference period, output from fct_combine_mcs_P()}

\item{.sim_mon}{simulation aggregated to monitoring period(s), output from fct_combine_mcs_P()}

\item{.ad_annual}{TRUE or FALSE, is the activity data annualized or not.}
}
\value{
A tibble with simulations at the final estimate per type of period.
}
\description{
Depending on how the period is defined and if the data are annualized
or not, calculate the Emission Level for a reference or monitoring period
for each simulation.
}
\examples{
path <- system.file("extdata/mocaredd-templatev2-simple.xlsx", package = "mocaredd.dev2")

checked <- fct_checkinput(.path = path)

sim_trans <- fct_combine_mcs_E(.checked_data = checked)

sim_REF <- fct_combine_mcs_P(.data = sim_trans, .period_type = "REF")

sim_MON <- fct_combine_mcs_P(.data = sim_trans, .period_type = "MON")

## !!! SIM MON and ER to be done

}
