% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_E.R
\name{fct_combine_mcs_E}
\alias{fct_combine_mcs_E}
\title{Generate and propagate Monte Carlo Simulations based on a template input file.}
\usage{
fct_combine_mcs_E(.checked_data)
}
\arguments{
\item{.checked_data}{The list returned by \code{fct_checkinput()}. Its
\code{template_version} element drives version-specific behaviour and its
\code{data} element supplies the \code{setup}, \code{time}, \code{area}
and \code{carbon} tables.}
}
\value{
A data frame with Monte Carlo simulations of CO2 emissions for each land use
transition, REDD+ activity or emission reductions level.
}
\description{
TBD
}
\examples{
library(mocaredd)

path <- system.file("extdata/mocaredd-templatev2-simple.xlsx", package = "mocaredd.dev2")

checked <- fct_checkinput(.path = path)

res <- fct_combine_mcs_E(.checked_data = checked)

get_trans <- sample(res$trans_id, 1)
res_sub <- res |> dplyr::filter(trans_id == get_trans)

hist(res_sub$E)

}
